# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Response:
    properties:
      data: {}
      status:
        $ref: '#/definitions/model.status'
    type: object
  model.status:
    enum:
    - success
    - error
    type: string
    x-enum-varnames:
    - success
    - err
  pagination.Pagination:
    properties:
      current_page:
        example: 1
        type: integer
      per_page:
        example: 1
        type: integer
      total:
        example: 10
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  post.PostResponse:
    properties:
      age:
        type: integer
      animal_type:
        type: string
      author_username:
        type: string
      color:
        type: string
      comments:
        type: integer
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      gender:
        type: string
      is_favourite:
        type: boolean
      photo:
        items:
          type: integer
        type: array
      status:
        type: string
      title:
        type: string
    type: object
  post.Response:
    properties:
      meta:
        $ref: '#/definitions/pagination.Pagination'
      posts:
        items:
          $ref: '#/definitions/post.PostResponse'
        type: array
    type: object
  role.DeleteRole:
    properties:
      name:
        enum:
        - keeper
        - seeker
        - vet
        type: string
    required:
    - name
    type: object
  role.GivenRole:
    properties:
      description:
        maxLength: 512
        type: string
      name:
        enum:
        - keeper
        - seeker
        - vet
        type: string
    required:
    - name
    type: object
  role.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  role.UpdateRole:
    properties:
      description:
        maxLength: 512
        type: string
      name:
        enum:
        - keeper
        - seeker
        - vet
        type: string
    required:
    - name
    type: object
...
  user.ResponseUser:
    properties:
      description:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      photo:
        items:
          type: integer
        type: array
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API for the SOS Kotopes project. It provides endpoints
    for managing the database of animals in need of help.
  title: SOS Kotopes API
  version: "1.0"
paths:
...
  /posts/{id}/favourites:
    delete:
      consumes:
      - application/json
      description: Delete post from favourites
      operationId: delete-favourite-post
      parameters:
      - description: Post ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Delete post from favourites
      tags:
      - post
    post:
      consumes:
      - application/json
      description: Add post to favourites
      operationId: add-favourite-post
      parameters:
      - description: Post ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Add post to favourites
      tags:
      - post
...
  /posts/favourites:
    get:
      consumes:
      - application/json
      description: Get favourite posts of user by id
      operationId: get-favourite-posts-user-by-id
      parameters:
      - description: Limit
        in: query
        minimum: 1
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        minimum: 0
        name: offset
        required: true
        type: integer
      - description: Status
        in: query
        name: status
        type: string
      - description: Animal type
        in: query
        name: animal_type
        type: string
      - description: Gender
        in: query
        name: gender
        type: string
      - description: Color
        in: query
        name: color
        type: string
      - description: Location
        in: query
        name: location
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/post.Response'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Get favourite posts of user by id
      tags:
      - post
...
  /users:
    patch:
      consumes:
      - application/json
      description: Update user
      operationId: update-user
      parameters:
      - description: Username
        in: formData
        name: username
        type: string
      - description: First name
        in: formData
        name: firstname
        type: string
      - description: Last name
        in: formData
        name: lastname
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Photo
        in: formData
        name: photo
        type: file
      - description: Password
        in: formData
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.ResponseUser'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update user
      tags:
      - user
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user by id
      operationId: get-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.ResponseUser'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Get user by id
      tags:
      - user
  /users/{id}/roles:
    get:
      consumes:
      - application/json
      description: Get user roles
      operationId: get-user-roles
      parameters:
      - description: User ID
        in: path
        name: request
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/role.Role'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get user roles
      tags:
      - role
  /users/roles:
    delete:
      consumes:
      - application/json
      description: Delete user role
      operationId: delete-user-role
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.DeleteRole'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Delete user role
      tags:
      - role
    patch:
      consumes:
      - application/json
      description: Update user roles
      operationId: update-user-roles
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.UpdateRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/role.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Update user roles
      tags:
      - role
    post:
      consumes:
      - application/json
      description: Give role to user
      operationId: give-role-to-user
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.GivenRole'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/role.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthBasic: []
      summary: Give role to user
      tags:
      - role
securityDefinitions:
  ApiKeyAuthBasic:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"